pyproject.toml
[project]
scripts = { beancount-import = "beancount_importer_rules.cli:cli" }


name = "beancount-importer-rules"
version = "0.1.0"
description = "Beancount importer rules engine"
authors = [
    { name = "Zeno Jiricek", email = "airtonix@users.noreply.github.com" },
]
dependencies = [
    "beancount-parser>=1.2.3",
    "beancount-black>=1.0.4",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.2",
    "pydantic>=2.8.2",
    "pytz>=2024.1",
    "click>=8.1.7",
    "rich>=13.8.0",
    "pendulum>=3.0.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.3.2",
    "pytest-mock>=3.14.0",
    "rich>=13.8.0",
    "beancount==2.3.6",
    "pylance>=0.16.1",
]

[tool.pdm.build]
includes = [
    "beancount_importer_rules",
    "README.md",
    "pyproject.toml",
    "LICENSE",
]

[tool.pdm.scripts]
start = "flask run -p 54321"
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
